cloudfx/3/requirements.txt
tweepy==3.8.0
pandas==1.0.3  
google-cloud-storage
datetime
pandas_gbq